# zstd: Improve memory usage on small streaming encodes