--- a/huff0/_generate/go.mod
+++ b/huff0/_generate/go.mod
@@ -1,6 +1,8 @@
 module github.com/klauspost/compress/s2/_generate
 
-go 1.19
+go 1.21
+
+toolchain go1.22.4
 
 require (
 	github.com/klauspost/compress v1.15.15