--- a/s2/cmd/_s2sx/go.mod
+++ b/s2/cmd/_s2sx/go.mod
@@ -1,6 +1,8 @@
 module github.com/klauspost/compress/s2/cmd/s2sx
 
-go 1.19
+go 1.21
+
+toolchain go1.22.4
 
 require github.com/klauspost/compress v1.11.9